refactor: staged_ledger_diff.rs with transaction_status.rs
