+++ protocol/test-serialization/src/lib.rs
@@ -16,6 +16,7 @@ mod tests {
     use super::{block_path_test, block_path_test_batch};
     use bin_prot::{from_reader, to_writer, Value};
     use mina_crypto::{hash::*, prelude::*, signature::*};
+    use mina_rs_base::transaction_status::*;
     use mina_rs_base::types::{
         proof_messages::{ProofMessageWithDegreeBound, ProofMessageWithoutDegreeBoundList},
         *,
