+++ protocol/test-serialization/src/test_3NKaBJsN1SehD6iJwRwJSFmVzJg5DXSUQVgnMxtH4eer4aF5BrDK.rs
@@ -5,6 +5,7 @@
 mod tests {
     use anyhow::bail;
     use mina_crypto::prelude::*;
+    use mina_rs_base::transaction_status::*;
     use mina_rs_base::types::*;
     use pretty_assertions::assert_eq;
     use test_fixtures::*;
